pyproject.toml
[tool.poetry]
name = "blog_project"
version = "0.1.0"
description = ""
authors = ["Tim Vogt <tvogt@intothesource.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.2"
django-cors-headers = "^4.0.0"
python-dotenv = "^1.1.0"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
dj-database-url = "^2.3.0"
gunicorn = "^23.0.0"
whitenoise = "^6.9.0"
Pillow = "^10.4.0"
django-cleanup = "^8.1.0"
django-storages = "^1.14.4"
boto3 = "^1.34.0"
cloudinary = "^1.40.0"
django-cloudinary-storage = "^0.3.0"

[tool.poetry.group.dev.dependencies]
# hier kun je dev-tools zetten zoals pytest of black


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
